Add events-form js controller
